requirements.txt
# Основные зависимости для оптимизированного бота
python-telegram-bot==13.15
python-dotenv==1.0.0
requests==2.31.0

# Опциональные зависимости
# httpx[socks]==0.25.0  # Для OpenAI с прокси
# openai==1.3.0         # Для OpenAI API
